+++ src/app/components/SessionOtherResources/index.tsx
@@ -15,6 +15,7 @@ import darkMode from "@/app/images/versions/1_21_0/dark-mode.png";
 import SaraIa from "@/app/images/versions/1_22_0/sara-ia.png";
 import Communication from "@/app/images/versions/1_23_0/communication.png";
 import IaProblema from "@/app/images/versions/1_24_0/problem-ai.png";
+import Indicator from "@/app/images/versions/1_25_0/indicator.png";
 
 export function SessionOtherResources() {
   return (
@@ -22,6 +23,14 @@ export function SessionOtherResources() {
       <div className="dark:bg-slate-900 bg-slate-100 md:p-10 p-4 rounded-2xl space-y-10">
         <h2 className="text-3xl text-center">Conheça outros recursos</h2>
         <div className="grid xl:grid-cols-3 gap-10 md:grid-cols-2">
+          <OtherVersion
+            link="/versao-1-25-0"
+            date="27 de fevereiro de 2025"
+            imageUrl={Indicator}
+            version="v1.25.0"
+            title="Indicadores"
+          />
+
           <OtherVersion
             link="/versao-1-24-0"
             date="13 de janeiro de 2025"
